+++ src/XrdClHttp/XrdClHttpFileSystemPlugIn.cc
@@ -77,8 +77,9 @@ XRootDStatus HttpFileSystemPlugIn::Mv(const std::string &source,
                             + dest;
 
   logger_->Debug(kLogXrdClHttp,
-                 "HttpFileSystemPlugIn::Mv - src = %s, dest = %s, timeout = %d",
-                 full_source_path.c_str(), full_dest_path.c_str(), timeout);
+                 "HttpFileSystemPlugIn::Mv - src = %s, dest = %s, timeout = %lld",
+                 full_source_path.c_str(), full_dest_path.c_str(),
+                 (long long)timeout);
 
   auto status =
       Posix::Rename(*davix_client_, full_source_path, full_dest_path, timeout);
@@ -100,8 +101,8 @@ XRootDStatus HttpFileSystemPlugIn::Rm(const std::string &path,
   url.SetPath(path);
 
   logger_->Debug(kLogXrdClHttp,
-                 "HttpFileSystemPlugIn::Rm - path = %s, timeout = %d",
-                 url.GetObfuscatedURL().c_str(), timeout);
+                 "HttpFileSystemPlugIn::Rm - path = %s, timeout = %lld",
+                 url.GetObfuscatedURL().c_str(), (long long)timeout);
 
   auto status = Posix::Unlink(*davix_client_, url.GetURL(), timeout);
 
@@ -125,8 +126,8 @@ XRootDStatus HttpFileSystemPlugIn::MkDir(const std::string &path,
 
   logger_->Debug(
       kLogXrdClHttp,
-      "HttpFileSystemPlugIn::MkDir - path = %s, flags = %d, timeout = %d",
-      url.GetObfuscatedURL().c_str(), flags, timeout);
+      "HttpFileSystemPlugIn::MkDir - path = %s, flags = %d, timeout = %lld",
+      url.GetObfuscatedURL().c_str(), flags, (long long)timeout);
 
   auto status = Posix::MkDir(*davix_client_, url.GetURL(), flags, mode, timeout);
   if (status.IsError()) {
@@ -146,8 +147,8 @@ XRootDStatus HttpFileSystemPlugIn::RmDir(const std::string &path,
   url.SetPath(path);
 
   logger_->Debug(kLogXrdClHttp,
-                 "HttpFileSystemPlugIn::RmDir - path = %s, timeout = %d",
-                 url.GetObfuscatedURL().c_str(), timeout);
+                 "HttpFileSystemPlugIn::RmDir - path = %s, timeout = %lld",
+                 url.GetObfuscatedURL().c_str(), (long long)timeout);
 
   auto status = Posix::RmDir(*davix_client_, url.GetURL(), timeout);
   if (status.IsError()) {
@@ -169,8 +170,8 @@ XRootDStatus HttpFileSystemPlugIn::DirList(const std::string &path,
 
   logger_->Debug(
       kLogXrdClHttp,
-      "HttpFileSystemPlugIn::DirList - path = %s, flags = %d, timeout = %d",
-      full_path.c_str(), flags, timeout);
+      "HttpFileSystemPlugIn::DirList - path = %s, flags = %d, timeout = %lld",
+      full_path.c_str(), flags, (long long)timeout);
 
   const bool details = flags & DirListFlags::Stat;
   const bool recursive = flags & DirListFlags::Recursive;
@@ -200,8 +201,8 @@ XRootDStatus HttpFileSystemPlugIn::Stat(const std::string &path,
                          std::to_string(url_.GetPort()) + "/" + path;
 
   logger_->Debug(kLogXrdClHttp,
-                 "HttpFileSystemPlugIn::Stat - path = %s, timeout = %d",
-                 full_path.c_str(), timeout);
+                 "HttpFileSystemPlugIn::Stat - path = %s, timeout = %lld",
+                 full_path.c_str(), (long long)timeout);
 
   auto stat_info = new StatInfo();
   //XRootDStatus status;
